Tipos que podem receber nulo
